Add files via upload
12V BATTERY LEVEL INDICATOR CIRCUIT NEW

 

The circuit I designed is suitable for 12V batteries. You include the + and - ends of the battery by soldering them to the vias. If the red led lights up, it means your battery is low. If the yellow led lights up, your battery is medium charged. If the green led lights up, it means your battery is fully charged.

I designed it for you as a PCB that is very simple to solder and use. Thanks.

 

COMPONENTS

D1 : 8V Zener Diode

D2 : 9V Zener Diode

D3 : 10V Zener Diode

R1, R2, R3 : 670 ohm THT Resistor

D4 : 5mm Red Led

D5 : 5mm Yellow Led

D6 : 5mm Green Led



Zener diode is composed of P and N type semiconductor materials, it is made of silicon. They are diodes that keep the voltage applied to their ends constant. Zener diodes do not pass current up to a certain voltage value. This voltage is called the break or zener voltage (Vz). When connected to the circuit in the right direction, it works like a normal diode. When it is connected in the reverse direction, it does not conduct until the breakdown voltage (zener voltage), and when the breakdown voltage is exceeded, it passes current in the form of an avalanche effect and turns on. When the reverse voltage is removed, the zener diode returns to its normal state. Since Zener diodes work in reverse polarity, they are reverse connected to the circuit. The Zener point (breakdown voltage) value is determined by adjusting the amount of additive during the production phase.